+++ app/src/main/java/com/emotionstorage/emotionstorage/ui/AppNavHost.kt
@@ -34,7 +34,7 @@ import com.emotionstorage.time_capsule_detail.ui.SaveTimeCapsuleScreen
 import com.emotionstorage.time_capsule_detail.ui.TimeCapsuleDetailScreen
 import com.emotionstorage.tutorial.ui.OnBoardingNavHost
 import com.emotionstorage.tutorial.ui.SplashScreen
-import com.emotionstorage.tutorial.ui.tutorial.TutorialScreen
+import com.emotionstorage.tutorial.ui.TutorialScreen
 import com.emotionstorage.ui.component.appBar.BottomAppNavBar
 import com.emotionstorage.ui.component.appBar.BottomNavDest
 import com.emotionstorage.ui.theme.MooiTheme
